refactor : social image -> user image 이름변경 및 경로 alias 설정
